Task01/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scikit-image>=0.18.0
torch>=1.9.0
torchvision>=0.10.0
monai>=0.8.0
nibabel>=3.2.0
SimpleITK>=2.1.0
matplotlib>=3.4.0
seaborn>=0.11.0
jupyter>=1.0.0
pytest>=6.2.0
black>=21.5b2
flake8>=3.9.0
mypy>=0.910
pydantic>=1.8.0
pyyaml>=5.4.0
tqdm>=4.62.0 
wandb>=0.12.0
torchio>=0.20.0
pybids>=0.15.0
xgboost>=1.5.0
opencv-python>=4.5.0
tensorboard>=2.10.0
torchmetrics>=0.11.0